split provider config options from meter config
