+++ sdk/metric/x/configurator.go
@@ -4,10 +4,7 @@
 // Package x contains experimental SDK-level metric configuration.
 package x // import "go.opentelemetry.io/otel/sdk/metric/x"
 
-import (
-	"go.opentelemetry.io/otel/sdk/instrumentation"
-	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
-)
+import "go.opentelemetry.io/otel/sdk/instrumentation"
 
 // MeterConfig contains SDK runtime configuration for a Meter.
 // It is returned by [MeterConfigurator] and controls whether a Meter
@@ -54,25 +51,3 @@ func (o meterEnabledOption) applyMeterConfig(c *MeterConfig) {
 // for that Meter.
 type MeterConfigurator func(instrumentation.Scope) MeterConfig
 
-type meterConfiguratorProviderOption struct {
-	// nil embed, skip guard in newConfig and prevents apply from being called
-	sdkmetric.Option
-	configurator MeterConfigurator
-}
-
-// Experimental marks this as an experimental option so the skip guard
-// in newConfig() skips calling the nil embedded apply().
-func (meterConfiguratorProviderOption) Experimental() {}
-
-func WithMeterConfigurator(fn MeterConfigurator) sdkmetric.Option {
-	return meterConfiguratorProviderOption{configurator: fn}
-}
-
-// MeterConfigurator returns the configurator as func(scope) any so
-// sdk/metric can extract it via duck-type without importing this package.
-
-func (co meterConfiguratorProviderOption) MeterConfigurator() func(instrumentation.Scope) any {
-	return func(s instrumentation.Scope) any {
-		return co.configurator(s)
-	}
-}
